Tapping and better workspace switching
When using ctrl+a for tmux I would have to hold the key down for a while to not just trigger esc, halving the tapping term from 200 to 100 seems to have helped a lot.

Slowly utilising more workspaces in xmonad, to the point that having 1-5 on my left hand wasn't enough. Using U, I, and O to switch monitor focus means that I can't just use the same setup on the right.
Created a combo of f and j to one shot a new layer (5) on which the top row is LGUI(n) (with the same layout as 0-9). Will see how this goes, but makes switching to workspaces 6+ much nicer